Enable products for channels tests

+++ server/channels/store/sqlstore/store.go
@@ -1000,7 +1000,7 @@ func (ss *SqlStore) DropAllTables() {
 			    FROM   pg_class
 			    WHERE  relkind = 'r'  -- only tables
 			    AND    relnamespace = 'public'::regnamespace
-				AND NOT relname = 'db_migrations'
+				AND NOT (relname = 'db_migrations' OR relname = 'focalboard_schema_migrations')
 			   );
 			END
 			$func$;`)
@@ -1008,7 +1008,7 @@ func (ss *SqlStore) DropAllTables() {
 		tables := []string{}
 		ss.masterX.Select(&tables, `show tables`)
 		for _, t := range tables {
-			if t != "db_migrations" {
+			if t != "db_migrations" && t != "focalboard_schema_migrations" {
 				ss.masterX.Exec(`TRUNCATE TABLE ` + t)
 			}
 		}
